implemented token based authentication system
When user signs up, a spot releated to users given PK is set t NULL. On future logins, a auth token and issued at timestamp releated to the logged in user will be sat and updated into the table. On response to login we use this token to authenticate user and use token for all future user releated API requests. Token is stored in localstorage and removed on logout. If user has not checked of "Remember me", token is stored in sessionStorage and removed on browser/window exit.